Move `shader_manager.cpp|h` into `src/gui/render`
