+++ src/gui/private/sdl_gui.h
@@ -1,10 +1,8 @@
-// SPDX-FileCopyrightText:  2022-2025 The DOSBox Staging Team
+// SPDX-FileCopyrightText:  2022-2026 The DOSBox Staging Team
 // SPDX-License-Identifier: GPL-2.0-or-later
 
 #ifndef DOSBOX_SDLMAIN_H
 #define DOSBOX_SDLMAIN_H
-
-#include "gui/private/shader_manager.h"
 
 #include <cstring>
 #include <memory>
@@ -12,6 +10,8 @@
 #include <string>
 #include <string_view>
 #include <vector>
+
+#include "gui/render/private/shader_manager.h"
 
 #include "dosbox_config.h"
 #include "gui/common.h"
