+++ src/gui/render/private/shader_manager.h
